% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{check_na_rows}
\alias{check_na_rows}
\title{Check matrix to get the rows with NAs}
\usage{
check_na_rows(x)
}
\arguments{
\item{x}{Data Matrixm}
}
\value{
Vector, row numbers that contain NAs
}
\description{
Check matrix to get the rows with NAs
}
\examples{
na_row_idx <- check_na_rows(x)
}
