% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{DE_limma}
\alias{DE_limma}
\title{Differential expression analysis using limma approach}
\usage{
DE_limma(expr_matrix, species_list, log2_scale = TRUE)
}
\arguments{
\item{expr_matrix}{Matrix, with gene expression data. The rows are for genes, and columns for
samples. The samples should be from two conditions (ie species), as indicated
by species_list.}

\item{species_list}{List of species, which should match the columns of expr_matrix. There are
should be two species, each with at least 2 replicates.}

\item{log2_scale}{TRUE/FALSE, indicate if the expression data should be log2-scaled before use.}
}
\value{
Data frame, with p-values and other statistics from differential analysis.
}
\description{
Differential expression analysis using limma approach
}
\examples{
hs2mm.limma_result <- DE_limma(hs2mm.expr.norm, hs2mm.species_list)
}
