% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{DE_ttest}
\alias{DE_ttest}
\title{Differential expression analysis using t-test approach}
\usage{
DE_ttest(expr_matrix, species_list)
}
\arguments{
\item{expr_matrix}{Matrix, with gene expression data. The rows are for genes, and columns for
samples. The samples should be from two conditions (ie species), as indicated
by species_list.}

\item{species_list}{List of species, which should match the columns of expr_matrix. There are
should be two species, each with at least 2 replicates.}
}
\value{
Data frame, with p-values and other statistics from differential analysis.
}
\description{
Differential expression analysis using t-test approach
}
\examples{
hs2mm.tt_result <- DE_ttest(hs2mm.expr.norm, hs2mm.species_list)
}
