% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{expression_statistics}
\alias{expression_statistics}
\title{Calculate statistics from expression matrix}
\usage{
expression_statistics(expr_matrix, group_list)
}
\arguments{
\item{expr_matrix}{Matrix, with gene expression data. The rows are for genes, and columns for
samples. The samples should be from two conditions (ie species), as indicated
by species_list.}
}
\value{
Data frame, with basic statistics (eg. average expression level,
expression level for each species, log2foldChange etc.) calculated from
expression data matrix.
}
\description{
Based on the expression data matrix for differential expression analysis,
this function calculate basic statistics (eg. average expression level,
expression level for each species, log2foldChange etc.), which will be
returned together with differential analysis results (eg. P-values etc,
which are calculated else where).
}
\examples{
hs2mm.stat <- expression_statistics(hs2mm.expr.norm, hs2mm.species_list)
}
