% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{get_uniq_id_pair}
\alias{get_uniq_id_pair}
\title{Extract IDs that are unique in both lists}
\usage{
get_uniq_id_pair(x, y)
}
\arguments{
\item{x}{Vector, ID list 1}

\item{y}{Vector, ID list 2}
}
\description{
Based on two lists of IDs, this function check both of them and
return IDs that are unique in both lists. By applying to the
homolog annotation tabels of two species, it can can get
candidate 1-to-1 orthologue pairs which will be further
compared.
}
\examples{
get_uniq_id_pair(human.ortholog.ann$ensembl_gene_id, mouse.ortholog.ann$hsapiens_homolog_ensembl_gene)
get_uniq_id_pair(human.ortholog.ann[,1], mouse.ortholog.ann[,2])
}
