% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{list_species}
\alias{list_species}
\title{List all supported species}
\usage{
list_species()
}
\value{
Data frame, with the information for supported species
}
\description{
Only species with genome available in ENSEMBL database are supported by
ExprX. This function will show the information for all these species.
}
\examples{
list_species()

}
