% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{make_ExprX_dataset}
\alias{make_ExprX_dataset}
\title{Create ExprX dataset}
\usage{
make_ExprX_dataset(x, data_dir)
}
\arguments{
\item{x}{Data frame or CSV file, which contains the information for each expression
data file for different species.The function can automatically determine if
the given x is a data frame or file name. If it's a data frame, then use
directly. Otherwise, it will read the information by reading the CSV file.}

\item{data_dir}{The directory to find expression data files. If the absolute directory of
expression data files are specified, then this parameter won't be matter.
But if what is specified is relative directory (or no directory), then this
parameter will determine how to find the expression data files. Default: ./
(ie current working directory)}
}
\value{
ExprX object, with information for expression values for each sample of
each species. The created object will be used by most of the analysis,
including orthologue matching, data normalization and differential
expression analysis. After each operation, the results can be appended to
the original ExprX object.
}
\description{
By parsing the meta table (as a data frame or CSV file) which contains
information about expression data files (usually contain TPM, FPKM or RPKM
values) for different species, this function will read these data files to
create an object which contains the expression levels of the replicates of
different species. The created ExprX object can also contain additional data
such as orthologue pairs, normalized expression etc, and will be used by
most of the subsequent analysis.
}
\examples{
# Make an ExprX object based on manually created data frame which contains
the meta information
hs2mm.data <- make_ExprX_dataset(hsa2mmu_meta_info)

# make an ExprX object based on the CSV file with meta information
hs2mm.data <- make_ExprX_dataset("hsa2mmu_metatable.csv")

}
