% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{make_meta_table}
\alias{make_meta_table}
\title{Create meta table with information for expression data files of each species}
\usage{
make_meta_table(file_list, species_list, col_id, col_expr, expr_type)
}
\arguments{
\item{file_list}{Vector, list of expression data files.}

\item{species_list}{Vector, list of the species names (eg. human, mouse) which are in
corresponding to the expression data files. The list of all supported species
by ExprX can be obtained with the \code{\link{list_species}} function.}

\item{col_id}{Integer, the column with gene ID in the expression data files. For most cases
(eg. RSEM results), the Gene IDs are in the 1st column.}

\item{col_expr}{Integer, the column with expression values in the expression data files. For
example, the TPM values are in the 6th column of the RSEM result.}

\item{data_type}{Type of gene expression data. Should be: tpm, fpkm or rpkm.}
}
\value{
Data frame, with information for each expression data file for each species.
}
\description{
Based on given expression data file list, species list and information about
which columns have gene IDs and expression values, this function can create
a data frame with meta informaton for the expression data files for each
species.
}
\examples{
hs2mm.meta <- make_meta_table(
  file_list = c(
  "tpm_dir/hsa_tpm_1.txt", "tpm_dir/hsa_tpm_2.txt",
  "tpm_dir/mmu_tpm_1.txt", "tpm_dir/mmu_tpm_2.txt"
  ),
  species_list = rep(c("human", "mouse"), each = 2),
  col_id = 1, col_expr = 6, data_type = "tpm"
)

}
