% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_matrix}
\alias{normalize_matrix}
\title{Normalalize expression data matrix with specified method}
\usage{
normalize_matrix(x, method)
}
\arguments{
\item{x}{Expression data matrix, with rows for genes and columns for samples.}

\item{method}{Method for normalization. It can be TMM, TMMwsp, RLE, upperquantile, quantile
or none.}
}
\value{
Expression data matrix after normalization.
}
\description{
This function can normalize expression data matrix with specified approach.
Several approaches are implemented, including TMM, TMMwsp, RLE, upperquantile
and quantile. These approaches are implemented by invoking edgeR and limma.
}
\examples{
mat_tpm.nrm  <- normalize_matrix(mat_tpm, "TMM")
mat_fpkm.nrm <- normalize_matrix(mat_fpkm, "quantile")

}
