% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{ortholog_expression_compare}
\alias{ortholog_expression_compare}
\title{Differential expression analysis of 1:1 orthologs between species}
\usage{
ortholog_expression_compare(expr_data, method, p_adjust)
}
\arguments{
\item{expr_data}{ExprX object which contains normalized expression data. It can be generated
with \code{\link{expression_matrix_normalize}}.}

\item{method}{Method for differential expression analysis. Can be RankProd, limma, t-test.}

\item{p_adjust}{Method for p-value adjustment. Can be "holm", "hochberg", "hommel",
"bonferroni", "BH", "BY", "fdr", "none". P-value adjustment is
performed using the \code{p.adjust} function. Default: fdr.}
}
\value{
Dataframe, with basic information (eg. GeneID, GeneName) of 1:1 orthologs and
statistics (eg. log2foldChange, adjusted p-value) from differential analysis.
The returned matrix can be filtered based on P-value and log2foldChange to
get differential genes.
}
\description{
This function performs differential expression analysis of 1:1 orthologs
between species. Expression data of 1:1 orthologs after normalization are
used for differential expression analysis.Statistics such as average
expression level, log2foldChange and p-values are calculated and returned
as a dataframe.
}
\examples{
hs2mm.de <- ortholog_expression_compare(hs2mm.data, method = "RankProd", p_adjust = "fdr")

}
