% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{ortholog_expression_merge}
\alias{ortholog_expression_merge}
\title{Integrate the expression data of orthologues among species}
\usage{
ortholog_expression_merge(expr_data, orth_data)
}
\arguments{
\item{expr_data}{ExprX object. Contains the original expression data for the samples of
compared species. It can be generated with \code{\link{make_ExprX_dataset}}.}

\item{orth_data}{Ortholog data. Contains the ortholog matching results among compared
species. It can be generated with \code{\link{ortholog_match}}.}
}
\value{
Updated ExprX object. It has the same structure to expr_data, but with
the expression data matrix of 1:1 orthologs for all samples from compared
species appended.
}
\description{
Take the original ExprX object generated with \code{\link{make_ExprX_dataset}}
and the orthologue pairing data generated using \code{\link{ortholog_match}}
as input, this function integrates together the expression data for all 1:1
orthologs among compared species. The integrated data will be appended to
the original ExprX object and returned as an updated object. To be noted, for
orthologs that don't have matched expression data, they will be excluded from
the returned data.
}
\examples{
# update the ExprX object to include expression data matrix for matched
# 1:1 orthologs. The updated object can then be used for data normalization
# and differential analysis
hs2mm.data <- ortholog_expression_merge(hs2mm.data, hs2mm.orth)

}
