% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{ortholog_expression_normalize}
\alias{ortholog_expression_normalize}
\title{Normalize the expression data of orthologs among species.}
\usage{
ortholog_expression_normalize(expr_data, method)
}
\arguments{
\item{expr_data}{ExprX object. Contains the expression data for the 1:1 orthologs for
all samples of compared species. It can be generated using
\code{\link{ortholog_expression_merge}}.}

\item{method}{Normalization method. Supported methods include: TMM, TMMwsp, RLE,
upperquartile, quantile, none.}
}
\value{
Updated ExprX object. It is of the same structure of expr_data, but with
normalized expression data appended.
}
\description{
Take ExprX object as input, it normalizes the expression data for 1:1
orthologs among samples of compared species. Different normalization
approaches are supported, including TMM, TMMwsp, RLE, upperquartile and
quantile. The implement of these approaches are based on edgeR and limma.
The normalized data matrix will be appended to the original ExprX object and
returned as the updated ExprX object.
}
\examples{
# update the ExprX dataset to included normalized expression data matrix
# for 1:1 orthologs. The updated object can then be used for differential
# expression analysis
hs2mm.data <- ortholog_expression_normalize(hs2mm.data, method = "TMM")

}
