% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ortholog_expression_plot}
\alias{ortholog_expression_plot}
\title{Visualize differential expression analysis result}
\usage{
ortholog_expression_plot(
  d,
  type,
  cutoff_pvalue = 0.05,
  cutoff_log2fold = 1,
  main,
  xlab,
  ylab,
  xlim,
  ylim
)
}
\arguments{
\item{d}{Dataframe for differential analysis results. It can be generated with the
function \code{\link{ortholog_expression_compare}}.}

\item{type}{The type of visualization. MA or volcano. Default: MA}

\item{cutoff_pvalue}{The cutoff of p-value for differential genes. Set as 1 to keep all genes.
Default: 0.05.}

\item{cutoff_log2fold}{The cutoff of log2fold for determine differential genes. Set as 0 to
keep all genes. Default: 1.}

\item{main}{Main title, default: NULL}

\item{xlab}{xlabel, default: log2fold(species1/species2)}

\item{ylab}{ylabel, default: "Average expression" for MA-plot, and "-log10(p-value)"
for Volcano-plot}

\item{xlim}{Range of x-axis, default: NULL}

\item{ylim}{Range of y-axis, default: NULL}
}
\description{
Visualize differential analysis results of the 1:1 orthologous genes
between two species. The differential genes are defined based on the
given cutoffs of p-value and log2fold. The differential expression can be
visualized as MA-plot or Volcano-plot, with differential genes
highlighted by color.
}
\examples{

# Visualize differential analysis result as MA-plot, with genes with
# pvalue<0.05 highlighted
ortholog_expression_plot(hs2mm.de, type = "MA", cutoff_pvalue = 0.05)

# Visualize differential analysis result as Valcano-plot, with genes with
# pvalue<0.05 and abs(log2fold)>1 highlighted
ortholog_expression_plot(
  hs2mm.de, type = "volcano",
  cutoff_pvalue = 0.05, cutoff_log2fold = 1
)

}
