% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortholog.R
\name{ortholog_filter}
\alias{ortholog_filter}
\title{Filter 1-to-1 orthologous pairs based on genetype or chromosomes}
\usage{
ortholog_filter(
  x,
  genetype_include,
  genetype_exclude,
  chrom_include,
  chrom_exclude,
  geneid_include,
  geneid_exclude
)
}
\arguments{
\item{x}{Data with information for paired 1-to-1 orthologous pairs. It should be the
list generated using \code{\link{ortholog_match}}.}

\item{genetype_include}{Gene types to be included. Use \code{\link{summarize_ortholog_gene}} to get
full list. genetype_include and genetype_exclude are exclusive.}

\item{genetype_exclude}{Gene types to be excluded. Use \code{\link{summarize_ortholog_gene}} to get
full list.  genetype_include and genetype_exclude are exclusive.}

\item{chrom_include}{Chromosomes to be included. Use \code{\link{summarize_ortholog_gene}} to
get full list. chrom_include and chrom_exclude are exclusive.}

\item{chrom_exclude}{Chromosomes to be excluded. Use \code{\link{summarize_ortholog_gene}} to
get full list. chrom_include and chrom_exclude are exclusive.}

\item{geneid_include}{Ensembl gene IDs to be included. geneid_include and geneid_exclude are
exclusive.}

\item{geneid_exclude}{Ensembl gene IDs to be excluded. geneid_include and geneid_exclude are
exclusive.}
}
\value{
List, filtered data with the same structre as the input data d.
}
\description{
Take the result generated using ortholog_match() as input, and perform
filtering based on gene type (eg. protein_coding), chromosome name, or
gene list. It will return the filtered data with the same data structure.
}
\examples{
hs2mm.orth <- ortholog_match("human", "mouse")

hs2mm.orth.flt1 <- ortholog_filter(hs2mm.orth, genetype_include = "protein_coding")
hs2mm.orth.flt2 <- ortholog_filter(hsa2mmu.df, chrom_exclude = c("X", "Y", "M"))
hs2mm.orth.flt3 <- ortholog_filter(
  hs2mm.orth,
  genetype_include = "protein_coding",
  chrom_exclude    = c("X", "Y", "M"),
  geneid_exclude   = c("ENSG00000000003", "ENSG00000000005", "ENSG00000000419")
)

}
