% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Visualize differential analysis result as volcano plot}
\usage{
plot_volcano(d, cutoff_pvalue, cutoff_log2fold, main, xlab, ylab, xlim, ylim)
}
\arguments{
\item{d}{Dataframe for differential analysis results. It can be generated with the
function \code{\link{ortholog_expression_compare}}.}

\item{cutoff_pvalue}{The cutoff of p-value for differential genes. Set as 1 to keep all genes.
Default: 0.05.}

\item{cutoff_log2fold}{The cutoff of log2fold for determine differential genes. Set as 0 to
keep all genes. Default: 1.}

\item{main}{Main title, default: NULL}

\item{xlab}{xlabel, default: log2foldChange}

\item{ylab}{ylabel, default: -log10(p-value)}

\item{xlim}{Range of x-axis, default: NULL}

\item{ylim}{Range of y-axis, default: NULL}
}
\description{
Visualize differential analysis result as volcano plot
}
\examples{
plot_volcano(deg, cutoff_pvalue = 0.05, cutoff_log2fold = 1)
}
