% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{read_data_matrix_from_file}
\alias{read_data_matrix_from_file}
\title{Read data files to create an expression matrix}
\usage{
read_data_matrix_from_file(
  file_list,
  col_geneid,
  col_value,
  header = TRUE,
  geneid_list
)
}
\arguments{
\item{file_list}{Vector, a list of expression data files}

\item{col_geneid}{Integer, indicate which column has gene IDs}

\item{col_value}{Integer, indicate which column has expression values}

\item{header}{TRUE/FALSE, indicate if the files have header line. Default: TRUE.}

\item{geneid_list}{Vector, a list of gene IDs. It is optional. If given, genes absent from the
list will be excluded. Otherwise, all genes read from the expression data
files will be kept.}
}
\value{
Matrix, with expression values for each gene in each expression data file.
For the created matrix, each row is for a gene and each column for a sample.
}
\description{
Read all given expression data files to get gene IDs and expression values,
then merge them to create a expression matrix.
}
\examples{
filelist <- c("file1.tpm.txt", "file2.tpm.txt")

genelist <- c(
  "ENSG00000000003", "ENSG00000000005", "ENSG00000000419",
  "ENSG00000000457", "ENSG00000000460")

expr.all  <- read_data_matrix_from_file(filelist, 1, 6)
expr.part <- read_data_matrix_from_file(filelist, 1, 6, genelist)
}
