% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{read_geneid_from_file}
\alias{read_geneid_from_file}
\title{Read expression data files to get gene IDs}
\usage{
read_geneid_from_file(file_list, col_geneid, header = TRUE)
}
\arguments{
\item{file_list}{Vector, a list of expression data files}

\item{col_geneid}{Integer, indicate which column has gene IDs}

\item{header}{TRUE/FALSE, indicate if the files have header line. Default: TRUE.}
}
\value{
Vector with the full list of gene names.
}
\description{
Read the given expression data files to get the gene IDs. If the gene IDs
are consistent among the files, return the them. Otherwise alert and exit.
}
\examples{
filelist <- c("file1.tpm.txt", "file2.tpm.txt")
geneid_list <- read_geneid_from_file(filelist, col_geneid = 1)
}
