% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{read_meta_table}
\alias{read_meta_table}
\title{Read meta information from file}
\usage{
read_meta_table(filename)
}
\arguments{
\item{filename}{File name, for the CSV file created manually or by using the
\code{\link{save_meta_table}} function.}
}
\value{
Data frame, with information about expression data files for each species.
}
\description{
It reads the meta information from the file (CSV format) created manually or
with the \code{\link{save_meta_table}} function. Alternatively, the meta
information can be created directly with the \code{\link{make_meta_table}}
function.
}
\examples{
hs2mm.meta <- read_meta_table("hs2mm.meta_table.csv")

}
