% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{save_meta_table}
\alias{save_meta_table}
\title{Write data frame with meta information as a CSV file}
\usage{
save_meta_table(x, filename, overwrite = FALSE)
}
\arguments{
\item{x}{Data frame, with information about expression data files for each
species. It can be created with \code{\link{make_meta_table}} function.}

\item{filename}{The name of the file to store the meta data.}

\item{overwrite}{TRUE/FALSE, to overwrite an existed file or not.
Default: FALSE}
}
\description{
Write data frame with meta information as a CSV file
}
\examples{
save_meta_table(hs2mm.meta, "hs2mm.metafile.csv")

}
