% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortholog.R
\name{summarize_ortholog_gene}
\alias{summarize_ortholog_gene}
\title{Summarize the orthologues for each genetype or chromosome}
\usage{
summarize_ortholog_gene(x, group = c("genetype", "chrom"))
}
\arguments{
\item{x}{Data frame with ortholog annotations for two species. It is generated
using ortholog_match(species1, species2).}

\item{group}{genetype or chrom, the number of genes for each group based on specified
tags will be listed.}
}
\value{
Data frame with the count of genes for each genetype or chrom in each species.
}
\description{
Summarize the orthologues for each genetype or chromosome
}
\examples{
hs2mm.orth <- ortholog_match("human", "mouse")
summarize_ortholog_gene(hs2mm.orth, "genetype")

}
