% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{valid_ensembl_geneid}
\alias{valid_ensembl_geneid}
\title{Check if the Ensembl gene IDs are valid}
\usage{
valid_ensembl_geneid(id, species)
}
\arguments{
\item{id}{Vector. List of ENSEMBL gene IDs}

\item{species}{Species name. Use list_species() to list all species.}
}
\value{
Logical, return TRUE if all valid, otherwise alert and return FALSE
}
\description{
Check if the Ensembl gene IDs are valid
}
\examples{
valid_ensembl_geneid(hsa.id, "human")
}
