% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{verify_ExprX_dataset}
\alias{verify_ExprX_dataset}
\title{Examine if the ExprX dataset is valid}
\usage{
verify_ExprX_dataset(x)
}
\arguments{
\item{x}{ExprX object generated using \code{\link{make_ExprX_dataset}}.}
}
\value{
ExprX object with the same structure of input x
}
\description{
Examine if the given ExprX object is valid, and print warnings if missing
some required content.
}
\examples{
hs2mm.data <- verify_ExprX_dataset(hs2mm.data)
}
