% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{verify_group_list}
\alias{verify_group_list}
\title{Verify if the group list is normal}
\usage{
verify_group_list(group_list)
}
\arguments{
\item{x}{Vector, with class labels.}
}
\value{
No return values. Stop with error if failed.
}
\description{
The input can be things like group classes as numbers, or species classes
as characters and so on. This function will check if there are multiple
groups, and each group should have >=2 replicates. If failed, it will exit
with errors.
}
\examples{
verify_group_list(c(1,1,1,0,0))
verify_group_list(c("human", "human", "human", "mouse", "mouse"))
}
