% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{verify_matrix}
\alias{verify_matrix}
\title{Verify if the object is a valid matrix}
\usage{
verify_matrix(x)
}
\arguments{
\item{x}{Data Matrixm}
}
\value{
No return values. Stop with error if failed.
}
\description{
Verify if the object is a valid matrix
}
\examples{
verify_matrix(mat)
}
